DF-8053: Cast meter to float
We calculate the meter as a short.  Cast it to float before sending.

+++ example/main.c
@@ -138,7 +138,7 @@ static void packet_rx(struct waveform_t* waveform,
    waveform_send_data_packet(waveform, null_samples,
                              get_packet_len(packet), SPEAKER_DATA);
 
-   waveform_meter_set_float_value(waveform, "junk-snr", ctx->snr);
+   waveform_meter_set_float_value(waveform, "junk-snr", (float) ctx->snr);
    waveform_meters_send(waveform);
    ctx->snr = (ctx->snr + 1) % 1024;
 }
